requirements.txt
numpy==1.24.3                  # 데이터 전처리 및 .npy 파일 처리
opencv-python==4.9.0.80        # 카메라 캡처, 이미지 표시, 프레임 처리
mediapipe==0.10.3              # 손 랜드마크 추적
torch==2.2.1                   # LSTM 모델 학습 및 추론
scikit-learn==1.3.2            # LabelEncoder, train_test_split 등
customtkinter==5.2.1           # GUI 인터페이스 구성 (버튼, 창 등)
Pillow==10.2.0                 # 말풍선 렌더링 및 이미지 처리 (PIL)
